background-remover/requirement.txt
rembg>=2.0.0
pillow>=9.0.0
numpy>=1.20.0
